+++ Unreal/Plugins/AirSim/Source/UnrealSensors/UnrealWifiSensor.h
@@ -33,6 +33,7 @@ protected:
 	//virtual void setPointCloud(const msr::airlib::Pose& sensor_pose, msr::airlib::vector<msr::airlib::real_T>& point_cloud, msr::airlib::TTimePoint time_stamp) override;
 
 	void updateWifiRays();
+	TArray<msr::airlib::Pose> getBeaconActors();
 private:
 	using Vector3r = msr::airlib::Vector3r;
 	using VectorMath = msr::airlib::VectorMath;
@@ -48,7 +49,7 @@ private:
 	std::vector<float> wifiTraceMaxDistances; // in meter
 
 	void sampleSphereCap(int num_points, float opening_angle);
-	int traceDirection(FVector trace_start_position, FVector trace_end_position, TArray<msr::airlib::WifiHit> *WifiHitLogfloat, float traceRayCurrentDistance, float traceRayCurrentbounces, float traceRayCurrentSignalStrength, bool drawDebug);
+	int traceDirection(FVector trace_start_position, FVector trace_end_position, TArray<msr::airlib::WifiHit> *WifiHitLogfloat, float traceRayCurrentDistance, float traceRayCurrentbounces, float traceRayCurrentSignalStrength, bool drawDebug, FVector trace_origin);
 	void bounceTrace(FVector &trace_start_position, FVector &trace_direction, float &trace_length, const FHitResult &trace_hit_result, float &total_distance, float &signal_attenuation);
 	float angleBetweenVectors(FVector vector1, FVector vector2);
 	float getFreeSpaceLoss(float previous_distance, float added_distance);
@@ -63,8 +64,7 @@ private:
 	float traceRayMaxDistance; // Meter
 	float traceRayMaxBounces;
 	float traceRayMinSignalStrength;
-
-	TArray<AActor*> beacon_actors;
+	TArray<msr::airlib::Pose> beacon_poses;
 
 	int maxWifiHits = 5;